+++ CHANGELOG.md
@@ -20,6 +20,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - **37_industry** add process-based steel model as alternative to CES-tree branch
 - **47_regipol** add support for delaying quantity targets and improving regional emission tax convergence
 - **core** change of preference parameters and associated computation of interest rates/mark ups 	
+- **scripts** add script to automatically check project summations from piamInterfaces
+    [[#1587](https://github.com/remindmodel/remind/pull/1587)]
 
 ### fixed
 - fixed weights of energy carriers in `pm_IndstCO2Captured`